work-around compiler bug in ActiveRemoteClient
It seems that 2.9.1 has a bug when a constructor argument is passed to
superclass which makes a public field out of it under the same name:
scalac got confused and generated invokespecial byte-code which is
rejected by the verifier. One could argue that the verifier is too
strict in this case ...

So, I changed the names of the constructor arguments by prepending an
underscore and thus avoiding the clash.