=str - 19149 - Corrects the behavior of FlowOps.sliding
Generatively attempts to verify that it behaves exactly like the
collections library

Removes the use of Vector.builder for an easier-to-follow Vector
implementation while waiting for this to be reimplemented as a
GraphStage.

Also, the Builder implementation relies on undefined behavior:

"Builder.result() — Produces a collection from the added elements.
The builder's contents are undefined after this operation." -
http://www.scala-lang.org/api/current/#scala.collection.mutable.Builder